interpolate improve code

+++ fps_interpolate.py
@@ -51,17 +51,22 @@ source_codec = determine_codec(src)
 
 if source_codec == 'mjpeg':
     cmd = '''
-    ffmpeg -hide_banner -loglevel error -i {src} -y -filter:v "minterpolate='mi_mode=mci:mc_mode=aobmc:vsbmc=1:fps=60'" -c:v libx264 -crf 0 -preset ultrafast  {temp_mp4}
+    ffmpeg -hide_banner -loglevel error -i {src} -y -filter:v "minterpolate='mi_mode=mci:mc_mode=aobmc:vsbmc=1:fps={fps}'" -c:v libx264 -crf 0 -preset ultrafast  {temp_mp4}
     
 ffmpeg -hide_banner -loglevel error -i {temp_mp4} -vf scale=-1:720 -c:v libvpx-vp9 -b:v 0 -crf 27 -row-mt 1 -pix_fmt yuv420p -c:a libopus  -movflags use_metadata_tags {dst}
+
+'''    
+    cmd_not_working = '''
+    ffmpeg -hide_banner -loglevel error -i {src} -y -filter:v "minterpolate='mi_mode=mci:mc_mode=aobmc:vsbmc=1:fps={fps}'"  -filter:v scale=-1:720 -c:v libvpx-vp9 -b:v 0 -crf 27 -row-mt 1 -pix_fmt yuv420p -c:a libopus  -movflags use_metadata_tags {dst}
 
 '''
     cmd = cmd.format(src=src, dst=dst, fps=fps,temp_mp4=temp_mp4)
     print(cmd)
     os.system(cmd)
-    os.unlink(temp_mp4)
-    
-if source_codec == 'h265':
+    if os.path.exists(temp_mp4):os.unlink(temp_mp4)
+
+elif source_codec == 'h265':
+    quit('interpolation not implement')
     cmd = '''
 ffmpeg -hide_banner -loglevel error -i {src}  -y -map 0:v -c:v copy -bsf:v  hevc_mp4toannexb raw.h265
 	ffmpeg -hide_banner -loglevel error -fflags +genpts -r {fps}  -y -i raw.h265 -c:v copy {dst}
@@ -72,6 +77,7 @@ ffmpeg -hide_banner -loglevel error -i {src}  -y -map 0:v -c:v copy -bsf:v  hevc
     os.unlink('raw.h265')
     
 elif source_codec == 'h264':
+    quit('interpolation not implement')
     cmd = '''
     ffmpeg -hide_banner -loglevel error -i {src}  -y -map 0:v -c:v copy -bsf:v  h264_mp4toannexb  raw.h264
 	ffmpeg -fflags +genpts -r {fps}  -y -i raw.h264 -c:v copy {dst}
